backend/admin_panel/requirements.txt
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask==3.1.0
PyYAML==6.0.2
SQLAlchemy==2.0.40